+++ lib/tri/form/__init__.py
@@ -36,7 +36,7 @@ except ImportError:  # pragma: no cover
         return engines['django'].from_string(template_code)
 
 
-__version__ = '3.3.0'
+__version__ = '3.4.0'
 
 
 def capitalize(s):